server/requirements.txt
# AI/ML Libraries
google-generativeai
torch
torchaudio
transformers

# Text-to-Speech
sarvamai
pydub

# Environment Variables
python-dotenv

# HTTP Requests
requests

# Language Detection
langdetect

